backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-cloud-aiplatform>=1.74.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.19.0
firebase-admin>=6.5.0
google-genai>=1.0.0
chromadb>=1.0.0
google-adk>=0.3.0
pydantic>=2.10.0
python-dotenv>=1.0.1
httpx>=0.27.0
python-multipart>=0.0.18
slowapi>=0.1.9
